Split out renewal and ingress watching services
